requirements.txt
Django==1.10.2
django-cors-headers==1.2.2
django-extensions==1.7.4
django-filter==0.15.3
django-paypal==0.3.3
djangorestframework==3.4.7
pytz==2016.7
requests==2.11.1
six==1.10.0
pillow==3.4.2
